\name{predictsFunctions-package}
\alias{predictsFunctions-package}
\alias{predictsFunctions}
\docType{package}
\title{
\packageTitle{predictsFunctions}
}
\description{
\packageDescription{predictsFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{predictsFunctions}
\packageIndices{predictsFunctions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{predictsFunctions}

Maintainer: \packageMaintainer{predictsFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
